Fix compilation error in MavenITmng11133MixinsPrecedenceTest
The test was using verifier.loadFile() which returns List<String> but assigning it to a String variable. Changed to use Files.readString() to read the file content as a String, which matches the expected variable type.

This fixes the compilation error:
incompatible types: java.lang.String cannot be converted to java.io.File