Bug: when raw-streams are used, ensure system streams are set up
When --raw-streams are used (especially when combined with options
like --quiet and -DforceStdout) there is no guarantee that
anything touches terminal. Hence, in case of embedded executor
it is quite possible that cached/warm code arrives fast at
the place that would do system out before the thread with
FastTerminal finishes.

In other words, when --raw-streams are used, we cannot right
now guarantee that system streams are already set up. This PR
changes that, and makes sure by triggering a dummy call
to terminal, at the cost of "jline3 install lag" for CLI
invocation. OTOH, this lag in case of embedded executors
does not exists (it exists only on first invocation).