make classworlds.conf configurable in CLI
This change allows setting the classworlds.conf property from the command line.
Currently, the only way to provide a custom Plexus extension is to patch the Maven distribution — either by modifying m2.conf or by adding a library into the lib directory. There is no other way to run Maven from the CLI with a custom workspace resolution.
With this change, it becomes possible to run Maven from the command line with a custom m2.conf file.