Add command line information to Verifier log files for easier IT reproduction
When running integration tests, the Verifier dumps output to a log file (usually log.txt).
This change adds the full command line used for execution at the beginning of the log file
to make it easy to reproduce the test outside the IT environment.

The command line information includes:
- Full path to Maven executable with all arguments
- Environment variables (MAVEN_OPTS, MAVEN_SKIP_RC, custom env vars)
- JVM system properties
- Working directory
- Execution mode (EMBEDDED, FORKED, AUTO)

The information is prepended to the log file after Maven execution completes,
ensuring it appears at the top without being overwritten by Maven's logging.

Example output:
# Command line used for this execution:
/opt/maven/bin/mvn -l log.txt --errors --batch-mode validate
# MAVEN_OPTS=-Duser.home=/path/to/user.home
# Working directory: /path/to/test/project
# Execution mode: AUTO