Revert the quoting of filenames in `JavaPathType.option(…)`. (#11435)
This is a partial revert of https://github.com/apache/maven/pull/2505
based on the observation that above PR has not been merged in 4.0.x.
For making the two branches consistent, we need to either port or revert 2505.
A revert is less disruptive as no Maven 4.1.x version has been released yet
and it would avoid https://github.com/apache/maven-compiler-plugin/pull/991.